Cache

Description

Cache typeclass definition.

Abstract

class (Show key, Monad m, Functor m) => Cache cache m key v where

An interface for a cache.

Minimal complete definition

inCache, listCache, lookupCache, filterCacheByKey, updateCache, insertInCache, updateOrInsertM

Methods

inCache :: cache key v -> key -> m Bool

Exists in cache?

listCache :: cache key v -> m [(key, v)]

Lists all cache entries.

lookupCache :: cache key v -> key -> m (Maybe v)

Search for a cached value by given key.

findInCacheWithDefault :: cache key v -> v -> key -> m v

Search for a cached value by given key, returning the default value if none was found.

filterCacheByKey :: cache key v -> (key -> Bool) -> m [(key, v)]

updateCache :: cache key v -> key -> (v -> v) -> m (cache key v)

Update a cached value (found by key) using the provided function.

insertInCache :: cache key v -> key -> v -> m (cache key v)

Insert a key -> value entry in the cache.

findOrElseInsert :: cache key v -> v -> key -> m (v, cache key v)

Find a value by key or insert the default value otherwise.

findOrElseInsertM :: cache key v -> m v -> key -> m (v, cache key v)

Find a value by key or insert the default value otherwise.

updateOrInsert :: cache key v -> (Maybe v -> v) -> key -> m (cache key v)

Update a cache entry or insert new if none exists.

updateOrInsertM :: cache key v -> (Maybe v -> m v) -> key -> m (cache key v)

Update a cache entry or insert new if none exists.

IO Cache

data IOCache key val

IO Cache.

newIOCache :: IO (IOCache key val)

Creates an empty IOCache.