EventProbability

Description

Definitions for events and probabilities.

Events

newtype Event

An event: intersection of underlying EvAtoms.

Constructors

Event (Map EventName EvAtom)

newtype EventName

A name of an event.

Constructors

EventName String

mkEvent :: EvAtom -> Event

Creates a singleton Event.

emptyEvent :: Event

Creates an empty Event.

compositeEvent :: (AtomicEvent e, Typeable e, Show e, Ord e) => Set e -> Event

Creates an intersection Event.

Atomic Events

class AtomicEvent e where

Atomic (not composite) event.

Methods

eventName :: e -> EventName

Event's name.

eventDomain :: e -> Set e

Event's values domain.

data EvAtom

An existential container for AtomicEvent.

Constructors

forall e . (AtomicEvent e, Show e, Ord e, Typeable e) => EvAtom e

class EventIntersection e1 e2 where

Events intersection builder.

Minimal complete definition

intersect

Methods

intersect :: e1 -> e2 -> Event

Events intersection.

(&) :: e1 -> e2 -> Event

Alias for intersect.

Probability

data Prob

Probability value container.

mkProb :: Double -> Prob

Create a Prob.

getProb :: Prob -> Double

Get Double value from Prob.

class Monad m => ProbabilityEstimation context m where

Interface for Events probability estimation.

Methods

estimateProb :: context -> Event -> m Prob

Estimate the probability of an event, given a context.

estimateCondProb

Arguments

:: context

-> Event
the event

-> EvAtom
condition

-> m Prob

Estimate the the conditional probability of an event, given a context.