EventProbability.Cache

Description

Count Cache

type EventCountCache cache m = Cache cache m Event Int

A cache for counting Events.

countOccurences :: EventCountCache cache m => cache Event Int -> Event -> m Int

given an event, searches in cache for events, that contain the event in question, and returns thier sum.

updateCountCache :: EventCountCache cache IO => cache Event Int -> [Event] -> IO (cache Event Int)

updates events count (accumulative) in cache.

Prob Caches

type EventProbCache cache m = Cache cache m Event Prob

A cache for Events probabilities.

type EventCondProbCache cache m = Cache cache m (Event, EvAtom) Prob

A cache for Events conditional probabilities.

Containers

data EventCaches m

A container for EventProbCache, EventProbCache and EventCondProbCache.

Constructors

forall cc pc cpc . (EventCountCache cc m, EventProbCache pc m, EventCondProbCache cpc m) => EvCaches

Fields

countCache :: cc Event Int

probCache :: pc Event Prob

condProbCache :: cpc (Event, EvAtom) Prob