EventProbability.NaiveBayes.Weka

Description

NaiveBayes learing and classification for Weka data.

Prepare

wekaEntry2Event :: Show WekaVal => WekaEntry -> Event

Build an Event from a WekaEntry.

Execute

buildCaches :: Show WekaVal => RawWekaData -> IO (EventCaches IO)

Create initial EventCaches and update counts of the EventCountCache.

classifyWekaData

Arguments

:: (Show WekaVal, NaiveBayesCondProb (EventCaches IO) IO)

=> EventCaches IO

-> RawWekaData
data to classify.

-> String
attribute to clasify.

-> IO [ClassifyResult]

Classifies given RawWekaData by given attribute name using EventCaches.

type ClassifyResult = (Event, Maybe (EvAtom, Prob))

Classification result of an Event.

Test

learnAndTest

Arguments

:: (Show WekaVal, Show WekaEntry, NaiveBayesCondProb (EventCaches IO) IO)

=> RawWekaData
learn data.

-> RawWekaData
test data.

-> String
attribute to classify.

-> Bool
show test results.

-> IO (EventCaches IO, [ClassifyResult])

Teaches a Naive Bayes classifier (updates the caches) with the learn data and tests the classification quality with the test data.

type Filename = String

runWekaLearnAndTest

Arguments

:: (Show WekaVal, Show WekaEntry, NaiveBayesCondProb (EventCaches IO) IO)

=> Filename
path to a weka data file with learn data.

-> Filename
path to a weka data file with test data.

-> String
attribute to classify.

-> Bool
show test results.

-> IO (EventCaches IO, [ClassifyResult])

Calls learnAndTest after extracting RawWekaData from the given files.